+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2024-06-26
+- add architecture test with `phpat`
+
 ## [1.3.0] - 2024-06-19
 - add `coinbase.com` as exchange rate provider
 - add chain for currency providers and retries